backend/requirements.txt
Django>=4.2
psycopg2-binary
gunicorn   
djangorestframework
drf-spectacular
django-cors-headers